+++ source/Octopus.Tentacle.Client/Retries/RpcCallNoRetriesHandler.cs
@@ -10,94 +10,88 @@ namespace Octopus.Tentacle.Client.Retries
             Func<CancellationToken, Task> action,
             bool abandonActionOnCancellation,
             TimeSpan abandonAfter,
-            CancellationToken cancellationToken)
+            CancellationToken ct)
         {
-            await ((Func<CancellationToken, Task>)(async ct =>
+            if (!abandonActionOnCancellation)
             {
-                if (!abandonActionOnCancellation)
+                await action(ct).ConfigureAwait(false);
+            }
+
+            using var abandonCancellationTokenSource = new CancellationTokenSource();
+            using (ct.Register(() =>
+                   {
+                       // Give the actionTask some time to cancel on it's own.
+                       // If it doesn't assume it does not co-operate with cancellationTokens and walk away.
+                       abandonCancellationTokenSource.TryCancelAfter(abandonAfter);
+                   }))
+            {
+                var abandonTask = abandonCancellationTokenSource.Token.AsTask();
+
+                try
                 {
-                    await action(ct).ConfigureAwait(false);
-                }
+                    var actionTask = action(ct);
+                    var completedTask = await Task.WhenAny(actionTask, abandonTask).ConfigureAwait(false);
+                    if (actionTask != completedTask)
+                    {
+                        actionTask.IgnoreUnobservedExceptions();
+                    }
 
-                using var abandonCancellationTokenSource = new CancellationTokenSource();
-                using (ct.Register(() =>
-                       {
-                           // Give the actionTask some time to cancel on it's own.
-                           // If it doesn't assume it does not co-operate with cancellationTokens and walk away.
-                           abandonCancellationTokenSource.TryCancelAfter(abandonAfter);
-                       }))
+                    await completedTask.ConfigureAwait(false);
+                }
+                catch (Exception e) when (e is OperationCanceledException)
                 {
-                    var abandonTask = abandonCancellationTokenSource.Token.AsTask();
-
-                    try
+                    if (abandonCancellationTokenSource.IsCancellationRequested)
                     {
-                        var actionTask = action(ct);
-                        var completedTask = await Task.WhenAny(actionTask, abandonTask).ConfigureAwait(false);
-                        if (actionTask != completedTask)
-                        {
-                            actionTask.IgnoreUnobservedExceptions();
-                        }
-
-                        await completedTask.ConfigureAwait(false);
+                        throw new OperationAbandonedException(e, abandonAfter);
                     }
-                    catch (Exception e) when (e is OperationCanceledException)
-                    {
-                        if (abandonCancellationTokenSource.IsCancellationRequested)
-                        {
-                            throw new OperationAbandonedException(e, abandonAfter);
-                        }
 
-                        throw;
-                    }
+                    throw;
                 }
-            }))(cancellationToken).ConfigureAwait(false);
+            }
         }
 
         public async Task<T> ExecuteWithNoRetries<T>(
             Func<CancellationToken, Task<T>> action,
             bool abandonActionOnCancellation,
             TimeSpan abandonAfter,
-            CancellationToken cancellationToken)
+            CancellationToken ct)
         {
-            return await ((Func<CancellationToken, Task<T>>)(async ct =>
+            if (!abandonActionOnCancellation)
             {
-                if (!abandonActionOnCancellation)
+                return await action(ct).ConfigureAwait(false);
+            }
+
+            using var abandonCancellationTokenSource = new CancellationTokenSource();
+            using (ct.Register(() =>
+                   {
+                       // Give the actionTask some time to cancel on it's own.
+                       // If it doesn't assume it does not co-operate with cancellationTokens and walk away.
+                       abandonCancellationTokenSource.TryCancelAfter(abandonAfter);
+                   }))
+            {
+                var abandonTask = abandonCancellationTokenSource.Token.AsTask<T>();
+
+                try
                 {
-                    return await action(ct).ConfigureAwait(false);
-                }
+                    var actionTask = action(ct);
+                    var completedTask = await Task.WhenAny(actionTask, abandonTask).ConfigureAwait(false);
+                    if (actionTask != completedTask)
+                    {
+                        actionTask.IgnoreUnobservedExceptions();
+                    }
 
-                using var abandonCancellationTokenSource = new CancellationTokenSource();
-                using (ct.Register(() =>
-                       {
-                           // Give the actionTask some time to cancel on it's own.
-                           // If it doesn't assume it does not co-operate with cancellationTokens and walk away.
-                           abandonCancellationTokenSource.TryCancelAfter(abandonAfter);
-                       }))
+                    return await completedTask.ConfigureAwait(false);
+                }
+                catch (Exception e) when (e is OperationCanceledException)
                 {
-                    var abandonTask = abandonCancellationTokenSource.Token.AsTask<T>();
-
-                    try
+                    if (abandonCancellationTokenSource.IsCancellationRequested)
                     {
-                        var actionTask = action(ct);
-                        var completedTask = await Task.WhenAny(actionTask, abandonTask).ConfigureAwait(false);
-                        if (actionTask != completedTask)
-                        {
-                            actionTask.IgnoreUnobservedExceptions();
-                        }
-
-                        return await completedTask.ConfigureAwait(false);
+                        throw new OperationAbandonedException(e, abandonAfter);
                     }
-                    catch (Exception e) when (e is OperationCanceledException)
-                    {
-                        if (abandonCancellationTokenSource.IsCancellationRequested)
-                        {
-                            throw new OperationAbandonedException(e, abandonAfter);
-                        }
 
-                        throw;
-                    }
+                    throw;
                 }
-            }))(cancellationToken).ConfigureAwait(false);
+            }
         }
     }
 }
