+++ source/Octopus.Tentacle.Tests.Integration/ClientScriptExecutionScriptServiceV1IsNotRetried.cs
@@ -6,11 +6,8 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Halibut;
 using NUnit.Framework;
-using Octopus.Tentacle.Client;
-using Octopus.Tentacle.Client.Scripts;
 using Octopus.Tentacle.CommonTestUtils.Builders;
 using Octopus.Tentacle.Contracts;
-using Octopus.Tentacle.Contracts.ScriptServiceV2;
 using Octopus.Tentacle.Tests.Integration.Support;
 using Octopus.Tentacle.Tests.Integration.Util;
 using Octopus.Tentacle.Tests.Integration.Util.Builders;
@@ -59,7 +56,7 @@ namespace Octopus.Tentacle.Tests.Integration
                 .Build();
 
             List<ProcessOutput> logs = new List<ProcessOutput>();
-            Assert.ThrowsAsync<HalibutClientException>(async () => await clientTentacle.TentacleClient.ExecuteScriptAssumingException(startScriptCommand, logs, CancellationToken));
+            Assert.ThrowsAsync<HalibutClientException>(async () => await clientTentacle.TentacleClient.ExecuteScript(startScriptCommand, logs, CancellationToken));
 
             // Let the script finish.
             File.WriteAllText(waitForFile, "");
@@ -74,7 +71,7 @@ namespace Octopus.Tentacle.Tests.Integration
 
             // We must ensure all script are complete, otherwise if we shutdown tentacle while running a script the build can hang.
             // Ensure the script is finished by running the script again, the isolation mutex will ensure this second script runs after the first is complete.
-            await clientTentacle.TentacleClient.ExecuteScriptAssumingException(startScriptCommand, new List<ProcessOutput>(), CancellationToken);
+            await clientTentacle.TentacleClient.ExecuteScript(startScriptCommand, new List<ProcessOutput>(), CancellationToken);
         }
 
         [Test]
@@ -115,7 +112,7 @@ namespace Octopus.Tentacle.Tests.Integration
 
             List<ProcessOutput> logs = new List<ProcessOutput>();
             Logger.Information("Starting and waiting for script exec");
-            Assert.ThrowsAsync<HalibutClientException>(async () => await clientTentacle.TentacleClient.ExecuteScriptAssumingException(startScriptCommand, logs, CancellationToken));
+            Assert.ThrowsAsync<HalibutClientException>(async () => await clientTentacle.TentacleClient.ExecuteScript(startScriptCommand, logs, CancellationToken));
             Logger.Information("Exception thrown.");
 
             // Let the script finish.
@@ -176,7 +173,7 @@ namespace Octopus.Tentacle.Tests.Integration
             List<ProcessOutput> logs = new List<ProcessOutput>();
 
             //Assert.CatchAsync(async () => await clientTentacle.TentacleClient.ExecuteScriptAssumingException(startScriptCommand, logs, cts.Token));
-            Assert.ThrowsAsync<HalibutClientException>(async () => await clientTentacle.TentacleClient.ExecuteScriptAssumingException(startScriptCommand, logs, cts.Token));
+            Assert.ThrowsAsync<HalibutClientException>(async () => await clientTentacle.TentacleClient.ExecuteScript(startScriptCommand, logs, cts.Token));
 
             // Let the script finish.
             File.WriteAllText(waitForFile, "");
@@ -220,7 +217,7 @@ namespace Octopus.Tentacle.Tests.Integration
             var startScriptCommand = new StartScriptCommandV2Builder().WithScriptBody(new ScriptBuilder().Print("hello")).Build();
 
             List<ProcessOutput> logs = new List<ProcessOutput>();
-            Assert.ThrowsAsync<HalibutClientException>(async () => await clientTentacle.TentacleClient.ExecuteScriptAssumingException(startScriptCommand, logs, CancellationToken));
+            Assert.ThrowsAsync<HalibutClientException>(async () => await clientTentacle.TentacleClient.ExecuteScript(startScriptCommand, logs, CancellationToken));
 
             // We Can not verify what will be in the logs because of race conditions in tentacle.
             // The last complete script which we fail might come back with the logs.
@@ -228,23 +225,6 @@ namespace Octopus.Tentacle.Tests.Integration
             scriptServiceExceptions.CompleteScriptLatestException.Should().NotBeNull();
             scriptServiceCallCounts.StartScriptCallCountStarted.Should().Be(1);
             scriptServiceCallCounts.CompleteScriptCallCountStarted.Should().Be(1);
-        }
-    }
-
-    static class TentacleClientExtensionMethods
-    {
-        public static async Task<ScriptExecutionResult> ExecuteScriptAssumingException(
-            this TentacleClient tentacleClient,
-            StartScriptCommandV2 startScriptCommand,
-            List<ProcessOutput> logs,
-            CancellationToken token)
-        {
-            var finalResponse = await tentacleClient.ExecuteScript(startScriptCommand,
-                onScriptStatusResponseReceived => logs.AddRange(onScriptStatusResponseReceived.Logs),
-                cts => Task.CompletedTask,
-                new SerilogLoggerBuilder().Build().ForContext<TentacleClient>().ToILog(),
-                token);
-            return (finalResponse);
         }
     }
 }
